radosgw: Remove the newline from swift errors.
The current code emits a newline after swift errors, but fails
to account for it when it calculates 'content-length'.  This results in
some clients (go github.com/ncw/swift) producing complaints about the
unsolicited newline such as this,
	Unsolicited response received on idle HTTP channel starting with "\n"; err=<nil>

This patch removes the newline.  That makes "go" happy, but it breaks
swift tests.  So this still needs more work to add that newline back in,
but somewhere else so it doesn't break error messages or http.

Signed-off-by: Marcus Watts <mwatts@redhat.com>